+++ kotlin-stdlib/common/main/kotlin/com/javiersc/kotlin/stdlib/Identity.kt
@@ -1,0 +1,3 @@
+package com.javiersc.kotlin.stdlib
+
+public inline fun <A> identity(a: A): A = a